refactor(web): migrate notes API interactions to OpenAPI client and update configurations
- Removed custom API utility functions and types for notes.
- Updated note types to use OpenAPI-generated types for improved type safety.
- Refactored the notes page and hooks to utilize the new OpenAPI client methods.
- Replaced ESLint configuration with Biome for linting and formatting.
- Added new TypeScript configuration for Next.js in next.config.ts.
- Updated package.json scripts for linting and formatting with Biome.
