+++ frontend/biome.json
@@ -1,7 +1,10 @@
 {
 	"$schema": "https://biomejs.dev/schemas/2.0.6/schema.json",
 	"vcs": { "enabled": false, "clientKind": "git", "useIgnoreFile": false },
-	"files": { "ignoreUnknown": false },
+	"files": {
+		"includes": ["**/*.ts", "**/*.tsx"],
+		"ignoreUnknown": true
+	},
 	"formatter": { "enabled": true, "indentStyle": "tab" },
 	"linter": {
 		"enabled": true,
