+++ frontend/hooks/useNote.ts
@@ -1,18 +1,23 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { notesApi } from "@/lib/api/notes";
-import type { Note, NoteCreate, NoteUpdate } from "@/types/note";
+import type { NoteCreate, NoteUpdate } from "@/lib/client";
+import {
+	createNoteApiNotesPost,
+	deleteNoteApiNotesNoteIdDelete,
+	getNotesApiNotesGet,
+	updateNoteApiNotesNoteIdPut,
+} from "@/lib/client/sdk.gen";
 
 export const useNotes = () =>
 	useQuery({
 		queryKey: ["notes"],
-		queryFn: notesApi.getAll,
+		queryFn: () => getNotesApiNotesGet().then((res) => res.data),
 		staleTime: 5 * 60 * 1000,
 	});
 
 export const useCreateNote = () => {
 	const queryClient = useQueryClient();
 	return useMutation({
-		mutationFn: (data: NoteCreate) => notesApi.create(data),
+		mutationFn: (data: NoteCreate) => createNoteApiNotesPost({ body: data }),
 		onSuccess: () => queryClient.invalidateQueries({ queryKey: ["notes"] }),
 	});
 };
@@ -21,7 +26,12 @@ export const useUpdateNote = () => {
 	const queryClient = useQueryClient();
 	return useMutation({
 		mutationFn: ({ id, data }: { id: number; data: NoteUpdate }) =>
-			notesApi.update(id, data),
+			updateNoteApiNotesNoteIdPut({
+				path: {
+					note_id: id,
+				},
+				body: data,
+			}),
 		onSuccess: () => queryClient.invalidateQueries({ queryKey: ["notes"] }),
 	});
 };
@@ -29,7 +39,12 @@ export const useUpdateNote = () => {
 export const useDeleteNote = () => {
 	const queryClient = useQueryClient();
 	return useMutation({
-		mutationFn: (id: number) => notesApi.delete(id),
+		mutationFn: (id: number) =>
+			deleteNoteApiNotesNoteIdDelete({
+				path: {
+					note_id: id,
+				},
+			}),
 		onSuccess: () => queryClient.invalidateQueries({ queryKey: ["notes"] }),
 	});
 };
